:sparkles: Implemented message object return & private commands
