+++ pincer/objects/__init__.py
@@ -22,8 +22,8 @@
 # TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 # SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-from .application import Application
 from .app_command import AppCommand
+from .application import Application
 from .attachment import Attachment
 from .audit_log import AuditLog
 from .ban import Ban
@@ -39,9 +39,9 @@ from .guild_widget import GuildWidget
 from .integration import Integration
 from .intents import Intents
 from .interaction_base import MessageInteraction
-from .interactions import Interaction
+from .interactions import Interaction, InteractionFlags
 from .invite import Invite
-from .message import Message
+from .message import Message, AllowedMentions, AllowedMentionTypes
 from .message_component import MessageComponent
 from .message_reference import MessageReference
 from .overwrite import Overwrite
@@ -52,6 +52,7 @@ from .stage import StageInstance
 from .sticker import Sticker
 from .thread import ThreadMember, ThreadMetadata
 from .user import User
+from .user_message import UserMessage
 from .voice_region import VoiceRegion
 from .voice_state import VoiceState
 from .webhook import Webhook
@@ -62,7 +63,9 @@ __all__ = (
     "Channel", "Embed", "Emoji", "FollowedChannel", "GuildMember",
     "GuildTemplate", "GuildWidget", "Guild", "Integration", "Interaction",
     "MessageInteraction", "Invite", "MessageComponent", "MessageReference",
-    "Message", "Overwrite", "Reaction", "Role", "SelectMenu", "SelectOption",
+    "UserMessage", "Overwrite", "Reaction", "Role", "SelectMenu",
+    "SelectOption", "Message",
     "StageInstance", "Sticker", "ThreadMember", "ThreadMetadata", "User",
-    "VoiceRegion", "VoiceState", "Webhook", "WelcomeScreen", "Intents"
+    "VoiceRegion", "VoiceState", "Webhook", "WelcomeScreen", "Intents",
+    "UserMessage", "AllowedMentions", "AllowedMentionTypes", "InteractionFlags"
 )
